# app/modules/profile/router.py
"""Profile router — GET/PATCH /profile/me, POST /profile/resume, GET /profile/completion-score."""
from fastapi import APIRouter, Depends, UploadFile, File

from app.modules.profile.service import ProfileService
from app.modules.profile.repository import ProfileRepository
from app.schemas.request.profile import ProfileUpdateRequest
from app.schemas.response.profile import (
    ProfileResponse, ParsedResumeResponse, CompletionScoreResponse
)
from app.shared.dependencies import get_db, get_current_user, get_structured_provider
from app.shared.response_models import ok, APIResponse

router = APIRouter(prefix="/profile", tags=["profile"])


def _get_service(db=Depends(get_db), llm=Depends(get_structured_provider)) -> ProfileService:
    return ProfileService(ProfileRepository(db), llm)


@router.get("/me", response_model=APIResponse[ProfileResponse])
async def get_profile(current_user: dict = Depends(get_current_user), service: ProfileService = Depends(_get_service)):
    return ok(data=service.get_profile(current_user["id"]))


@router.patch("/me", response_model=APIResponse[ProfileResponse])
async def update_profile(
    body: ProfileUpdateRequest,
    current_user: dict = Depends(get_current_user),
    service: ProfileService = Depends(_get_service),
):
    updated = service.update_profile(current_user["id"], body)
    return ok(data=updated, message="Profile updated.")


@router.post("/resume", response_model=APIResponse[ParsedResumeResponse])
async def upload_resume(
    resume: UploadFile = File(...),
    current_user: dict = Depends(get_current_user),
    service: ProfileService = Depends(_get_service),
):
    file_bytes = await resume.read()
    result = await service.upload_resume(
        user_id=current_user["id"],
        filename=resume.filename or "resume.pdf",
        content_type=resume.content_type or "",
        file_bytes=file_bytes,
    )
    return ok(data=result)


@router.get("/completion-score", response_model=APIResponse[CompletionScoreResponse])
async def get_completion(
    current_user: dict = Depends(get_current_user),
    service: ProfileService = Depends(_get_service),
):
    return ok(data=service.get_completion_score(current_user["id"]))