# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillBridge AI",
    "description": "AI-powered career guidance for India's workforce",
    "version": "1.0.0"
  },
  "paths": {
...
    "/profile/me": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_profile_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ProfileResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "profile"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_profile_me_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ProfileResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/profile/resume": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Upload Resume",
        "operationId": "upload_resume_profile_resume_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_profile_resume_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ParsedResumeResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/profile/completion-score": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Completion",
        "operationId": "get_completion_profile_completion_score_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_CompletionScoreResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/resume/analyze": {
      "post": {
        "tags": [
          "resume-analysis"
        ],
        "summary": "Analyze Resume",
        "description": "Analyzes a PDF resume and returns a detailed structured report.\nUses the production-grade multi-stage extraction pipeline.",
        "operationId": "analyze_resume_api_v1_resume_analyze_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_resume_api_v1_resume_analyze_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeAnalysisResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resume/analysis": {
      "get": {
        "tags": [
          "resume-analysis"
        ],
        "summary": "Get Latest Analysis",
        "description": "Returns the latest resume analysis for the current user.\nReturns None if no analysis exists.",
        "operationId": "get_latest_analysis_api_v1_resume_analysis_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ResumeAnalysisResult"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Latest Analysis Api V1 Resume Analysis Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resume/score-breakdown": {
      "get": {
        "tags": [
          "resume-analysis"
        ],
        "summary": "Get Score Breakdown",
        "description": "Returns only the scores and flags for dashboard widgets.\nReturns None if no analysis exists yet.",
        "operationId": "get_score_breakdown_api_v1_resume_score_breakdown_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resume/improve-bullet": {
      "post": {
        "tags": [
          "resume-analysis"
        ],
        "summary": "Improve Single Bullet",
        "description": "Improves a single bullet point. Rate limited.",
        "operationId": "improve_single_bullet_api_v1_resume_improve_bullet_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImproveBulletRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulletImprovement"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_CompletionScoreResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompletionScoreResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "APIResponse[CompletionScoreResponse]"
      },
...
      "APIResponse_ParsedResumeResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParsedResumeResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "APIResponse[ParsedResumeResponse]"
      },
      "APIResponse_ProfileResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfileResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "APIResponse[ProfileResponse]"
      },
...
      "Body_analyze_resume_api_v1_resume_analyze_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "target_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Role"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_resume_api_v1_resume_analyze_post"
      },
      "Body_upload_resume_profile_resume_post": {
        "properties": {
          "resume": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Resume"
          }
        },
        "type": "object",
        "required": [
          "resume"
        ],
        "title": "Body_upload_resume_profile_resume_post"
      },
...
      "BulletImprovement": {
        "properties": {
          "original": {
            "type": "string",
            "title": "Original"
          },
          "improved": {
            "type": "string",
            "title": "Improved"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "original",
          "improved",
          "reason"
        ],
        "title": "BulletImprovement"
      },
...
      "CompletionScoreResponse": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "filled_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filled Fields"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          }
        },
        "type": "object",
        "required": [
          "score",
          "filled_fields",
          "missing_fields"
        ],
        "title": "CompletionScoreResponse",
        "description": "Profile completeness analysis."
      },
...
      "EducationLevel": {
        "type": "string",
        "enum": [
          "none",
          "primary",
          "secondary",
          "higher_secondary",
          "graduate",
          "postgraduate",
          "vocational",
          "dropout",
          "10th",
          "12th"
        ],
        "title": "EducationLevel"
      },
...
      "Gender": {
        "type": "string",
        "enum": [
          "male",
          "female",
          "other",
          "prefer_not_to_say"
        ],
        "title": "Gender",
        "description": "Gender identification."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImproveBulletRequest": {
        "properties": {
          "bullet": {
            "type": "string",
            "title": "Bullet"
          },
          "target_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Role"
          }
        },
        "type": "object",
        "required": [
          "bullet"
        ],
        "title": "ImproveBulletRequest"
      },
...
      "Language": {
        "type": "string",
        "enum": [
          "hindi",
          "english",
          "bengali",
          "telugu",
          "marathi",
          "tamil",
          "urdu",
          "gujarati",
          "kannada",
          "odia",
          "punjabi",
          "malayalam",
          "hinglish"
        ],
        "title": "Language"
      },
...
      "ProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "education_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EducationLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "languages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Language"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Languages"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id"
        ],
        "title": "ProfileResponse",
        "description": "Detailed user profile information."
      },
      "ProfileUpdateRequest": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 14.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Region"
              },
              {
                "type": "null"
              }
            ]
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "education_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EducationLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "languages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Language"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Languages"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "title": "ProfileUpdateRequest",
        "description": "Generic profile update request.",
        "example": {
          "age": 22,
          "education_level": "graduate",
          "full_name": "Raj Kumar",
          "state": "Maharashtra"
        }
      },
...
      "Region": {
        "type": "string",
        "enum": [
          "andhra_pradesh",
          "arunachal_pradesh",
          "assam",
          "bihar",
          "chhattisgarh",
          "goa",
          "gujarat",
          "haryana",
          "himachal_pradesh",
          "jharkhand",
          "karnataka",
          "kerala",
          "madhya_pradesh",
          "maharashtra",
          "manipur",
          "meghalaya",
          "mizoram",
          "nagaland",
          "odisha",
          "punjab",
          "rajasthan",
          "sikkim",
          "tamil_nadu",
          "telangana",
          "tripura",
          "uttar_pradesh",
          "uttarakhand",
          "west_bengal",
          "delhi",
          "jammu_kashmir",
          "ladakh",
          "other"
        ],
        "title": "Region"
      },
...
      "ResumeAnalysisResult": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "structured_profile": {
            "$ref": "#/components/schemas/StructuredProfile"
          },
          "quality_scores": {
            "$ref": "#/components/schemas/QualityScores"
          },
          "suggestions": {
            "$ref": "#/components/schemas/SuggestionSet"
          },
          "overall_score": {
            "type": "integer",
            "title": "Overall Score",
            "default": 0
          },
          "target_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Role"
          },
          "india_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "India Flags"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "ResumeAnalysisResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}